requirement.txt
appdirs==1.4.4
arabic-reshaper==2.1.1
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==20.3.0
autopep8==1.4.4
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
blinker==1.4
boltons==20.2.1
bs4==0.0.1
certifi==2019.3.9
cffi==1.14.3
chardet==3.0.4
chatbot==1.5.2b0
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==7.1.1
colorama==0.4.4
comtypes==1.1.9
constants==0.6.0
contextlib2==0.6.0.post1
crispy-forms-gds==0.2.1
cryptography==3.2.1
cv==1.0.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.2.5
django-allauth==0.39.1
django-ckeditor==5.4.0
django-cleanup==2.1.0
django-cors-headers==3.7.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-email-extras==0.3.4
django-extensions==3.1.3
django-forms-builder==0.14.0
django-js-asset==1.0.0
django-simple-search==1.0.2
django-taggit==1.5.1
django-unused-media==0.2.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
docker==4.3.1
docopt==0.6.2
dump==0.0.5
emoji==0.6.0
entrypoints==0.3
face==20.1.1
filelock==3.0.12
Flask==1.1.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
future==0.15.0
gevent==20.9.0
google==3.0.0
greenlet==0.4.17
gunicorn==20.0.4
html5lib==1.1
humanize==3.10.0
idna==2.8
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
Js2Py==0.71
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
mathparse==0.1.2
matplotlib==3.3.3
mistune==0.8.4
MouseInfo==0.1.3
mysql-connector-python==8.0.22
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
nltk==3.5
notebook==6.1.5
numpy==1.19.4
oauthlib==3.0.1
opencv-python==4.5.1.48
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pep8==1.7.1
phpmyadmin==0.0.1
pickleshare==0.7.5
Pillow==6.2.2
Pint==0.16.1
pipwin==0.5.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.14.0
protocol==0.1.0
psycopg2==2.9.1
PyAudio @ file:///C:/Users/Danish%20khan/pipwin/PyAudio-0.2.11-cp38-cp38-win32.whl
PyAutoGUI==0.9.52
pycodestyle==2.5.0
pycparser==2.20
pycryptodome==3.10.1
pygame==2.0.0
PyGetWindow==0.0.9
Pygments==2.7.2
pyjokes==0.6.0
pyjsparser==2.7.1
PyJWT==2.1.0
pymongo==3.11.2
PyMsgBox==1.0.9
PyMySQL==0.10.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.1
pypiwin32==223
PyPrind==2.11.2
PyRect==0.1.4
pyrsistent==0.17.3
PyScreeze==0.1.26
pySmartDL==1.3.4
python-bidi==0.4.2
python-dateutil==2.7.5
python-decouple==3.1
python-gnupg==0.4.6
python-todo==0.4.1
python3-openid==3.1.0
pyttsx3==2.90
PyTweening==1.0.3
pytz==2018.5
pywhatkit==3.9
pywin32==227
pywinpty==0.5.7
PyYAML==3.13
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
redis==3.5.3
regex==2020.11.13
reportlab==3.5.59
requests==2.21.0
requests-oauthlib==1.2.0
schema==0.7.4
scikit-learn==0.23.2
scipy==1.5.4
selenium==3.141.0
Send2Trash==1.5.0
Shapely==1.7.1
six==1.15.0
sklearn==0.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
sphinx-me==0.3
sql==0.4.0
SQLAlchemy==1.3.23
sqlparse==0.2.4
stripe==2.27.0
termcolor==1.1.0
terminado==0.9.1
terminaltables==3.1.0
testpath==0.4.4
threadpoolctl==2.1.0
todo-cli==1.0.1
todo.py==0.3
todocli==3.2.1
tornado==6.1
tqdm==4.54.1
traitlets==5.0.5
tzlocal==2.1
Unidecode==1.2.0
urllib3==1.24.2
virtualenv==20.2.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
wikipedia==1.4.0
windows-curses==2.2.0
xhtml2pdf==0.2.5
zope.event==4.5.0
zope.interface==5.2.0
